+++ frost-ed25519/tests/integration_tests.rs
@@ -18,7 +18,7 @@ fn check_sign_with_dkg() {
 fn check_sign_with_cocktail_dkg() {
     let rng = rand::rngs::OsRng;
 
-    frost_core::tests::ciphersuite_generic::check_sign_with_cocktail_dkg::<Ed25519Sha512, _>(rng);
+    frost_core::tests::cocktail_dkg::check_sign_with_cocktail_dkg::<Ed25519Sha512, _>(rng);
 }
 
 #[test]
@@ -320,7 +320,7 @@ fn check_cocktail_dkg_test_vectors() {
 
     let json_str = include_str!("helpers/cocktail-dkg-ed25519-sha512.json");
 
-    frost_core::tests::ciphersuite_generic::check_cocktail_dkg_test_vectors::<Ed25519Sha512, _>(
+    frost_core::tests::cocktail_dkg::check_cocktail_dkg_test_vectors::<Ed25519Sha512, _>(
         json_str,
         |data| Sha512::digest(data).to_vec(),
         true, // encrypted shares match (XChaCha20Poly1305)
