+++ frost-ed448/tests/integration_tests.rs
@@ -18,7 +18,7 @@ fn check_sign_with_dkg() {
 fn check_sign_with_cocktail_dkg() {
     let rng = rand::rngs::OsRng;
 
-    frost_core::tests::ciphersuite_generic::check_sign_with_cocktail_dkg::<Ed448Shake256, _>(rng);
+    frost_core::tests::cocktail_dkg::check_sign_with_cocktail_dkg::<Ed448Shake256, _>(rng);
 }
 
 #[test]
@@ -323,7 +323,7 @@ fn check_cocktail_dkg_test_vectors() {
 
     let json_str = include_str!("helpers/cocktail-dkg-ed448-shake256.json");
 
-    frost_core::tests::ciphersuite_generic::check_cocktail_dkg_test_vectors::<Ed448Shake256, _>(
+    frost_core::tests::cocktail_dkg::check_cocktail_dkg_test_vectors::<Ed448Shake256, _>(
         json_str,
         |data| {
             let mut h = Shake256::default();
