+++ frost-p256/tests/integration_tests.rs
@@ -18,7 +18,7 @@ fn check_sign_with_dkg() {
 fn check_sign_with_cocktail_dkg() {
     let rng = rand::rngs::OsRng;
 
-    frost_core::tests::ciphersuite_generic::check_sign_with_cocktail_dkg::<P256Sha256, _>(rng);
+    frost_core::tests::cocktail_dkg::check_sign_with_cocktail_dkg::<P256Sha256, _>(rng);
 }
 
 #[test]
@@ -317,7 +317,7 @@ fn check_cocktail_dkg_test_vectors() {
 
     let json_str = include_str!("helpers/cocktail-dkg-p256-sha256.json");
 
-    frost_core::tests::ciphersuite_generic::check_cocktail_dkg_test_vectors::<P256Sha256, _>(
+    frost_core::tests::cocktail_dkg::check_cocktail_dkg_test_vectors::<P256Sha256, _>(
         json_str,
         |data| Sha256::digest(data).to_vec(),
         true, // encrypted shares match (XAES-256-GCM)
