+++ frost-secp256k1-tr/tests/integration_tests.rs
@@ -18,7 +18,7 @@ fn check_sign_with_dkg() {
 fn check_sign_with_cocktail_dkg() {
     let rng = rand::rngs::OsRng;
 
-    frost_core::tests::ciphersuite_generic::check_sign_with_cocktail_dkg::<Secp256K1Sha256TR, _>(
+    frost_core::tests::cocktail_dkg::check_sign_with_cocktail_dkg::<Secp256K1Sha256TR, _>(
         rng,
     );
 }
@@ -29,7 +29,7 @@ fn check_cocktail_dkg_test_vectors() {
 
     let json_str = include_str!("helpers/cocktail-dkg-secp256k1-sha256.json");
 
-    frost_core::tests::ciphersuite_generic::check_cocktail_dkg_test_vectors::<Secp256K1Sha256TR, _>(
+    frost_core::tests::cocktail_dkg::check_cocktail_dkg_test_vectors::<Secp256K1Sha256TR, _>(
         json_str,
         |data| Sha256::digest(data).to_vec(),
         true, // encrypted shares match (XAES-256-GCM)
